Changes to error propagation
Now generating an ensemble for each ensemble member and randomly
drawing one sample from that meta ensemble.  The idea here is to help
make sure we’re preserving variable covariances and allowing the means
to drift through time.

Previous versions were essentially doubling the anomaly so we weren’t
really getting true differences in climate due to the bias-correction.

** NOTE: Right now we are not adding in residual error.  We were
essentially doubling (or tripling) the anomalies and were getting a
20˚C range.  Right now Tmax looks like it gets a 3ish degree Celsius
range in temperature and this is from a combination of the error
associated with climate and anomaly bias correction only.