Temp code commit
Currently leaving in the myriad of alternative formulations I tried.
Including SWDOWN mostly helps.  Had to find a balance between which
predictors and interactions are important and removing anything not
essential to make sure we get decent uncertainties.  Current script is
applied at the daily time step and takes ~ 0.5 min/year to run.  There
will be a commit soon that cleans up this code.