initial swdown downscaling commit
currently filters by day, which is unnecessary since swdown always goes
to 0 at night.  This means it’s safe to model all years for a given DOY
at once. This will be done in the next commit.

this also means that swdown is more predictable than tair, so lets do
this first and then use it as an hourly predictor for tair.