save model & betas as .Rdata and .nc files
attempt to save memory & time; dunno if it’s actually going to work.

This required shifting the Rbeta calculation from predict.met to part
of the loop so we don’t have to load the full matrix.  Also added in
code to initialize the lags for the first time step from outside the
predict.subdaily function.

Also added a submit script to get these processing in the background
without my session timing out.