Fit models to moving window around the day of year
Some datasets like Ameriflux are too short and too noisy to be able to
train a truly day-specific model.  One reasonable way around this is to
fit the model a moving window with size of X.  For example, if we want
to model January 10, we can increase the amount of data used to train
this model by leveraging the diurnal pattern of the closest 5 days
(January 8-12).