Complete save of downscaling models and betas as separate files for each day
Now rather than running the model and generating a distribution of
betas for each day of the year at once, save each one as a separate
file to reduce memory demands.  Note, for high-performance machines
with lots of memory, this may be slower, but it will be more reliable
and flexible across platforms.