Load betas and model day by day rather than everything at once
Previously had saved models and betas for all days of year in a single
object.  However, this was very memory intensive and systems would
crash if not enough memory was available.  I’ve now switched that so
the betas and linear regression model for each day of year are saved as
separate objects, so we need to call them as such in this prediction
step.  This is probably going to be slower than saving everything in
one object, but it will be more flexible for low-memory computing or if
we want to implement sub-hourly models, which exponentially increase
the memory requirements.