fix pdsix bug; confirm moisture units must be INCHES
Bug in pdsix code came from initializing some variables with NA instead
of 0.  By comparing calculations with moisture units (precip, awc) in
mm instead of inches, found that although patterns are similar, it’s
not the same.  I might be overlooking a bug in the code though, so I’m
not going to rule it out, but it seems safer to work with English units
inside of the code.  Temperature doesn’t matter because it needs to be
in celsius for PE calculation.