Change to Precipitation Model
Adding a step to the precipitation model.  Similar to how we do in the
temporal downscaling model, we need to both correct the total annual
precipitation and the proportion of precipitation that falls at a given
time of year.  This looks like it solves the problem of negative precip
and will hopefully perform better with the GCMs.