Attempt to speed up code by pulling n betas from a random point.
Right now we’re reading in all of the Betas and then subsetting it in
memory.  It’s going to go a lot faster with a large set of coefficients
if we don’t read everything in first.  not *quite* as robust, but since
there’s already many levels of randomization and there should be no
correlation between row i and row i+1, it should be fine.